Impl [Project settings] Need an option to delete project from project settings
