+++ src/elements/ProjectCard/ProjectCardView.js
@@ -18,10 +18,12 @@ under the Apache 2.0 license is conditioned upon your compliance with
 such restriction.
 */
 import React, { useRef } from 'react'
+import { useSelector } from 'react-redux'
 import PropTypes from 'prop-types'
 import { useNavigate } from 'react-router-dom'
 
 import ActionsMenu from '../../common/ActionsMenu/ActionsMenu'
+import Loader from '../../common/Loader/Loader'
 import ReadOnlyChips from '../ReadOnlyChips/ReadOnlyChips'
 import ProjectStatistics from '../ProjectStatistics/ProjectStatistics'
 import { Tooltip, TextTooltipTemplate } from 'igz-controls/components'
@@ -37,72 +39,76 @@ const ProjectCardView = React.forwardRef(({ actionsMenu, project, statistics }, 
   const chipRef = useRef()
   const navigate = useNavigate()
 
-  return (
-    <div
-      className="project-card"
-      onClick={event => {
-        if (
-          event.target.tagName !== 'A' &&
-          !ref.current.contains(event.target) &&
-          !chipRef.current?.contains(event.target) &&
-          !event.target.closest('#overlay_container')
-        ) {
-          navigate(`/projects/${project.metadata.name}/monitor`)
-        }
-      }}
-      ref={cardRef}
-    >
-      <div className="project-card__general-info">
-        <div className="project-card__header">
-          <div className="project-card__header-title">
-            <Tooltip
-              className="project-card__title"
-              template={<TextTooltipTemplate text={project.metadata.name} />}
-            >
-              {project.metadata.name}
-            </Tooltip>
+  const { deletingProjects } = useSelector(state => state.projectStore)
 
-            <div className="project-card__info" data-testid="project-card__created">
-              <ClockIcon className="project-card__info-icon" />
-              <span>Created {getTimeElapsedByDate(project.metadata.created)}</span>
+  return (
+    <div className="project-card">
+      {Object.values(deletingProjects).includes(project.metadata.name) && <Loader />}
+      <div
+        onClick={event => {
+          if (
+            event.target.tagName !== 'A' &&
+            !ref.current.contains(event.target) &&
+            !chipRef.current?.contains(event.target) &&
+            !event.target.closest('#overlay_container')
+          ) {
+            navigate(`/projects/${project.metadata.name}/monitor`)
+          }
+        }}
+        ref={cardRef}
+      >
+        <div className="project-card__general-info">
+          <div className="project-card__header">
+            <div className="project-card__header-title">
+              <Tooltip
+                className="project-card__title"
+                template={<TextTooltipTemplate text={project.metadata.name} />}
+              >
+                {project.metadata.name}
+              </Tooltip>
+
+              <div className="project-card__info" data-testid="project-card__created">
+                <ClockIcon className="project-card__info-icon" />
+                <span>Created {getTimeElapsedByDate(project.metadata.created)}</span>
+              </div>
+            </div>
+
+            <div
+              className={`project-card__header-sub-title project-card__info ${
+                !project.spec.owner ? 'visibility-hidden' : ''
+              } `}
+              data-testid="project-card__owner"
+            >
+              <span>Owner:</span>
+              <span>{project.spec.owner}</span>
             </div>
           </div>
 
-          <div
-            className={`project-card__header-sub-title project-card__info ${
-              !project.spec.owner ? 'visibility-hidden' : ''
-            } `}
-            data-testid="project-card__owner"
-          >
-            <span>Owner:</span>
-            <span>{project.spec.owner}</span>
+          <div className="project-card__content">
+            <div className="project-card__description" data-testid="project-card__description">
+              {project?.spec.description && (
+                <Tooltip template={<TextTooltipTemplate text={project.spec.description} />}>
+                  {project.spec.description}
+                </Tooltip>
+              )}
+            </div>
+
+            <div className="project-card__statistic">
+              <ProjectStatistics statistics={statistics} />
+            </div>
           </div>
+
+          {project.metadata.labels && (
+            <div className="project-card__info" ref={chipRef} data-testid="project-card__labels">
+              <span>Labels:</span>
+              <ReadOnlyChips labels={project.metadata.labels} shortChips />
+            </div>
+          )}
         </div>
 
-        <div className="project-card__content">
-          <div className="project-card__description" data-testid="project-card__description">
-            {project?.spec.description && (
-              <Tooltip template={<TextTooltipTemplate text={project.spec.description} />}>
-                {project.spec.description}
-              </Tooltip>
-            )}
-          </div>
-
-          <div className="project-card__statistic">
-            <ProjectStatistics statistics={statistics} />
-          </div>
+        <div className="project-card__actions-menu" ref={ref}>
+          <ActionsMenu dataItem={project} menu={actionsMenu[project.metadata.name]} />
         </div>
-
-        {project.metadata.labels && (
-          <div className="project-card__info" ref={chipRef} data-testid="project-card__labels">
-            <span>Labels:</span>
-            <ReadOnlyChips labels={project.metadata.labels} shortChips />
-          </div>
-        )}
-      </div>
-
-      <div className="project-card__actions-menu" ref={ref}>
-        <ActionsMenu dataItem={project} menu={actionsMenu[project.metadata.name]} />
       </div>
     </div>
   )
