+++ src/elements/ProjectCard/projectCard.scss
@@ -93,6 +93,10 @@
     box-shadow: $previewBoxShadow;
   }
 
+  .loader-wrapper {
+    position: absolute;
+  }
+
   .project-data-card__statistics {
     &-item {
       flex: 0 0 auto;
